+++ TDEngine2/include/ecs/CWorld.h
@@ -100,6 +100,27 @@ namespace TDEngine2
 			TDE2_API E_RESULT_CODE DestroyImmediately(CEntity* pEntity) override;
 
 			/*!
+				\brief The method registers specified resource factory within a manager
+
+				\param[in] pFactory A pointer to IComponentFactory's implementation
+
+				\return The method returns an object, which contains a status of method's execution and
+				an identifier of the registred factory
+			*/
+
+			TDE2_API E_RESULT_CODE RegisterComponentFactory(TPtr<IComponentFactory> pFactory) override;
+
+			/*!
+				\brief The method unregisters a component factory with the specified identifier
+
+				\param[in] resourceFactoryId An identifier of a component factory
+
+				\return RC_OK if everything went ok, or some other code, which describes an error
+			*/
+
+			TDE2_API E_RESULT_CODE UnregisterComponentFactory(TypeId componentTypeId) override;
+
+			/*!
 				\brief The method registers specified system
 
 				\param[in] A pointer to ISystem's implementation