+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A new component which is **CSceneLoadingTriggerComponent** was added.
 
+- The API of **IWorld** was extended with new methods RegisterComponentFactory and UnregisterComponentFactory.
+
 ## [0.6.4] 2022-05-25
 
 ### Added